+++ spec/insert_output_snapshots/postgresql/insert-000-schema.sql
@@ -565,6 +565,16 @@ CREATE INDEX IF NOT EXISTS index_users_on_shop_id ON public.users USING btree (s
 
 
 --
+-- Name: users suppress_redundant_user_updates; Type: TRIGGER; Schema: public; Owner: -
+--
+
+DO $exwiw$ BEGIN
+  CREATE TRIGGER suppress_redundant_user_updates BEFORE UPDATE ON public.users FOR EACH ROW EXECUTE FUNCTION suppress_redundant_updates_trigger();
+EXCEPTION WHEN duplicate_object THEN NULL;
+END $exwiw$;
+
+
+--
 -- Name: transactions fk_rails_59d791a33f; Type: FK CONSTRAINT; Schema: public; Owner: -
 --